Renaming model to schema
