+++ app/parser/v0_0_1/schema_parser.py
@@ -9,16 +9,16 @@ from app.parser.abstract_schema_parser import AbstractSchemaParser
 from app.parser.schema_parser_exception import SchemaParserException
 from app.parser.parser_utils import ParserUtils
 
-from app.model.questionnaire import Questionnaire
-from app.model.group import Group
-from app.model.block import Block
-from app.model.section import Section
-from app.model.question import Question
-from app.model.answer import Answer
-from app.model.display import Display
-from app.model.properties import Properties
+from app.schema.questionnaire import Questionnaire
+from app.schema.group import Group
+from app.schema.block import Block
+from app.schema.section import Section
+from app.schema.question import Question
+from app.schema.answer import Answer
+from app.schema.display import Display
+from app.schema.properties import Properties
 
-from app.model.introduction import Introduction
+from app.schema.introduction import Introduction
 
 import logging
 import copy
