+++ app/state/answer.py
@@ -1,0 +1,7 @@
+
+
+class Answer(object):
+    def __init__(self, id):
+        self.id = id
+        self.is_valid = False
+        self.value = None
