+++ app/state/block.py
@@ -1,0 +1,7 @@
+
+
+class Block(object):
+
+    def __init__(self, id):
+        self.id = id
+        self.sections = []
