+++ app/state/group.py
@@ -1,0 +1,5 @@
+
+
+class Group(object):
+    def __init__(self, id):
+        self.id = id
