+++ app/state/question.py
@@ -1,0 +1,6 @@
+
+
+class Question(object):
+    def __init__(self, id):
+        self.id = id
+        self.answers = []
