+++ app/state/questionnaire.py
@@ -1,0 +1,6 @@
+
+
+class Questionnaire(object):
+    def __init__(self, id):
+        self.id = id
+        self.groups = []
