+++ app/state/section.py
@@ -1,0 +1,6 @@
+
+
+class Section(object):
+    def __init__(self, id):
+        self.id = id
+        self.questions = []
