+++ app/submitter/converter.py
@@ -50,7 +50,7 @@ class Converter(object):
         """
         Create the JSON answer format for down stream processing
 
-        :param questionnaire: the questionnaire model
+        :param questionnaire: the questionnaire schema
         :param answers: the users answers as a dict of id/value
         :return: a JSON object in the following format:
           {
