+++ app/surveys/large_components/groups/individual_questions.py
@@ -1,4 +1,4 @@
-from app.model.group import Group
+from app.schema.group import Group
 
 from app.surveys.large_components.pages.page7 import page7
 from app.surveys.large_components.pages.page8 import page8
