+++ app/surveys/large_components/pages/page8.py
@@ -1,7 +1,7 @@
-from app.model.block import Block
-from app.model.section import Section
-from app.model.question import Question
-from app.model.answer import Answer
+from app.schema.block import Block
+from app.schema.section import Section
+from app.schema.question import Question
+from app.schema.answer import Answer
 
 from app.surveys.large_test import test_questionnaire
