+++ tests/app/model/test_questionnaire.py
@@ -1,6 +1,6 @@
-from app.model.questionnaire import Questionnaire, QuestionnaireException
-from app.model.group import Group
-from app.model.block import Block
+from app.schema.questionnaire import Questionnaire, QuestionnaireException
+from app.schema.group import Group
+from app.schema.block import Block
 import unittest
