+++ tests/app/validation/test_type_validation_factory.py
@@ -2,7 +2,7 @@ import unittest
 from app.validation.type_validator_factory import TypeValidatorFactory, TypeValidatorFactoryException
 from app.validation.integer_type_check import IntegerTypeCheck
 from app.validation.date_type_check import DateTypeCheck
-from app.model.answer import Answer
+from app.schema.answer import Answer
 
 
 class TypeValidatorFactoryTest(unittest.TestCase):
